Move check_from_context to EventAuthHandler.
